requirements.txt
Django==3.0.8
django-mptt==0.11.0
django-ckeditor==5.6.1